JGH-OpenSourceLaw -
Took the Week description and i) renamed it by content instead of the artificial "/1/" for the week that it might (or might not) be in, ii) added a thesis as intro to the LessonPlan.

Renamed the "Bibliography.=" to "CourseBibliography.=" to avoid filling the "Bibliography.=" of each week (caused by deprefixing).